requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
aiogram>=3.4.0
openai>=1.14.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
apscheduler>=3.10.0
pydantic-settings>=2.2.0
boto3>=1.34.0
google-api-python-client>=2.120.0
google-auth>=2.28.0
httpx>=0.27.0
pdfplumber>=0.11.0
python-multipart>=0.0.9
